requirements.txt
# From pyproject.toml + backend
deepface>=0.0.98
python-dotenv>=1.0.0
nicegui>=3.7.1
qdrant-client>=1.16.2
sentence-transformers>=5.2.2
sentencepiece>=0.2.1
torch>=2.10.0
transformers>=5.1.0
Pillow>=10.0.0

# Backend (FastAPI)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
